# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stations": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get all stations",
                "operationId": "169c0910e38928ff56efb1027ff472fe",
                "responses": {
                    "200": {
                        "description": "List of all stations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Create a new station",
                "operationId": "6cecf862996dda377a5e3412fd5cf087",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Central Station"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Station created"
                    }
                }
            }
        },
        "/api/stations/search": {
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Search for stations by name prefix",
                "operationId": "ccf6f46b8c8393278cdc73a6e8835890",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starts_with"
                                ],
                                "properties": {
                                    "starts_with": {
                                        "type": "string",
                                        "example": "New"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of matching stations"
                    }
                }
            }
        },
        "/api/trains/search": {
            "post": {
                "tags": [
                    "Trains"
                ],
                "summary": "Search for trains between two stations",
                "operationId": "d9e2448f2d4fe13aab9774dafb8dec4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_station",
                                    "to_station",
                                    "departure_time"
                                ],
                                "properties": {
                                    "from_station": {
                                        "description": "Name of the departure station",
                                        "type": "string",
                                        "example": "Central Station"
                                    },
                                    "to_station": {
                                        "description": "Name of the destination station",
                                        "type": "string",
                                        "example": "North Station"
                                    },
                                    "departure_time": {
                                        "description": "Desired departure time in 'Y-m-d H:i:s' format",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-15 08:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of trains matching the search criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "trains": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}